// web/sturdy-potato/api-client/models/ProjectSchema.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Sturdy Potato API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface ProjectSchema
 */
export interface ProjectSchema {
    /**
     * 
     * @type {number}
     * @memberof ProjectSchema
     */
    id: number;
    /**
     * 
     * @type {string}
     * @memberof ProjectSchema
     */
    name: string;
    /**
     * 
     * @type {number}
     * @memberof ProjectSchema
     */
    taskCount?: number;
    /**
     * 
     * @type {string}
     * @memberof ProjectSchema
     */
    createdAt: string;
    /**
     * 
     * @type {string}
     * @memberof ProjectSchema
     */
    updatedAt: string;
}

/**
 * Check if a given object implements the ProjectSchema interface.
 */
export function instanceOfProjectSchema(value: object): value is ProjectSchema {
    if (!('id' in value) || value['id'] === undefined) return false;
    if (!('name' in value) || value['name'] === undefined) return false;
    if ((!('createdAt' in (value as Record<string, any>)) && !('created_at' in (value as Record<string, any>))) || ((value as Record<string, any>)['createdAt'] === undefined && (value as Record<string, any>)['created_at'] === undefined)) return false;
    if ((!('updatedAt' in (value as Record<string, any>)) && !('updated_at' in (value as Record<string, any>))) || ((value as Record<string, any>)['updatedAt'] === undefined && (value as Record<string, any>)['updated_at'] === undefined)) return false;
    return true;
}

export function ProjectSchemaFromJSON(json: any): ProjectSchema {
    return ProjectSchemaFromJSONTyped(json, false);
}

export function ProjectSchemaFromJSONTyped(json: any, ignoreDiscriminator: boolean): ProjectSchema {
    if (json == null) {
        return json;
    }
    return {
        
        'id': json['id'],
        'name': json['name'],
        'taskCount': json['task_count'] == null ? undefined : json['task_count'],
        'createdAt': json['created_at'],
        'updatedAt': json['updated_at'],
    };
}

export function ProjectSchemaToJSON(json: any): ProjectSchema {
    return ProjectSchemaToJSONTyped(json, false);
}

export function ProjectSchemaToJSONTyped(value?: ProjectSchema | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'id': value['id'],
        'name': value['name'],
        'task_count': value['taskCount'],
        'created_at': value['createdAt'],
        'updated_at': value['updatedAt'],
    };
}