// web/sturdy-potato/api-client/models/TodoSchema.ts
export interface TodoSchema {
    /**
     * 
     * @type {number}
     * @memberof TodoSchema
     */
    id: number;
    /**
     * 
     * @type {string}
     * @memberof TodoSchema
     */
    title: string;
    /**
     * 
     * @type {string}
     * @memberof TodoSchema
     */
    description: string;
    /**
     * 
     * @type {boolean}
     * @memberof TodoSchema
     */
    completed: boolean;
    /**
     * 
     * @type {number}
     * @memberof TodoSchema
     */
    projectId: number | null;
    /**
     * 
     * @type {Date}
     * @memberof TodoSchema
     */
    dueDate: Date | null;
    /**
     * 
     * @type {string}
     * @memberof TodoSchema
     */
    createdAt: string;
    /**
     * 
     * @type {string}
     * @memberof TodoSchema
     */
    updatedAt: string;
}